# [PERF] TSDB: Chunk encoding: shorten some write sequences