@beorn7 2024-09-18
I see, I guess it doesn't like the empty line 311, now that only closing braces are left.